Temporary workaround for MQTT ConcurrentSubscribeError and stability
Not fully clear about this yet but should at least help avoid
homeassistant-myskoda#140 and #123

When _something_ causes the MQTT connection to drop both the
`on_disconnect` and the `on_socket_close` fire and both trigger a
reconnect. This leads to a `ConcurrentSubscribeError` and all kinds of
weird networking and stablity issues throughout HA as a result.