Cache and update the last Vehicle (#369)
* Set timestamp in response models automatically

- Use field.default_factory to automatically set the timestamp field
when an instance is created
- Create a common base model to deduplicate the timestamp field

* Cache and update the last Vehicle

When intentionally exluding specific capabilities (~endpoints) the
problem is that the corresponding Vehicle fields will be None in the
returned Vehicle.

So either the client/caller must have the knowledge of the
capability->endpoint->field mapping and stitch together the previous and
new Vehicle, or the MySkoda class must store the last Vehicle(s) and
update what can be updated. The second option seems much cleaner and
keeps internal logic encapsulated in the myskoda lib.