BREAKING CHANGE: Refactor MySkoda and MQTT update logic
This moves all update logic out of the coordinator and into the MySkoda lib.

The coordinator can still subscribe to MQTT events but they are not used for anything and just get exposed in corresponding entities.

The coordinator can now subscribe to 'update' events which MySkoda will callback when any Vehicle data was updated for whatever reason (including when it, the MySkoda lib, got an MQTT event). When receiving an update callback the coordinator just has to call async_set_updated_data and pass in the current Vehicle data.