chore: remove amqtt, refactor affected tests
Fixes #401.

amqtt is old, unmaintained, and causes dependency issues.

Rather than trying to replace it we've refactored the tests and tested
code to no longer depend on a 'real' MQTT broker running. This focusses
the tests on our own code instead of mostly testing the MQTT client lib
(aiomqtt).

This is done by injecting a fake aiomqtt.Client in MySkodaMqttClient.