chore/refactor: remove amqtt, refactor tests + MySkodaMqttClient (#436)
* refactor: decouple aiomqtt.Client from MySkodaMqttClient

Allow tests injecting their own fake client so that we can test
MySkodaMqttClient without needing an actual MQTT broker and so we test
our own code and not the aiomqtt client code.

* chore: remove amqtt, refactor affected tests

Fixes #401.

amqtt is old, unmaintained, and causes dependency issues.

Rather than trying to replace it we've refactored the tests and tested
code to no longer depend on a 'real' MQTT broker running. This focusses
the tests on our own code instead of mostly testing the MQTT client lib
(aiomqtt).

This is done by injecting a fake aiomqtt.Client in MySkodaMqttClient.