feat: refresh charging on charging events
If the event is a charging event and not a change-soc event (which
already contains data itself) forece a refresh of the charging endpoint
data.

On Citigoe people seem to only receive `charging-status-changed` events,
which contain no usable data. Newer EVs (Enyaq, ...) don't see that
event at all and instead get `change-soc`.

Fixes: skodaconnect/homeassistant-myskoda#821