fix: update login flow, response_type changes
The login flow has been updated by VW/Skoda.

Including `id_token` in the `response_type` results in a
400/"invalid client" response.

The final redirect then just returns the code, which ultimate is all
we need anyway, so `_enter_password` can just return that and we no
longer need the `IDKAuthorizationCode` model.

Fixes #486 and skodaconnect/homeassistant-myskoda#932